Move RegisterProviderPath to abstract_provider.py
