+++ resources/lib/youtube_plugin/kodion/abstract_provider.py
@@ -322,3 +322,16 @@ class AbstractProvider(object):
 
     def tear_down(self, context):
         pass
+
+
+class RegisterProviderPath(object):
+    def __init__(self, re_path):
+        self._kodion_re_path = re_path
+
+    def __call__(self, func):
+        def wrapper(*args, **kwargs):
+            # only use a wrapper if you need extra code to be run here
+            return func(*args, **kwargs)
+
+        wrapper.kodion_re_path = self._kodion_re_path
+        return wrapper
